+++ web/src/lib/components/elements/buttons/circle-icon-button.svelte
@@ -14,7 +14,8 @@
 	{title}
 	style:background-color={backgroundColor}
 	style:--immich-icon-button-hover-color={hoverColor}
-	class="dark:text-immich-dark-fg rounded-full p-3 flex place-items-center place-content-center transition-all
+	class:dark:text-immich-dark-fg={!forceDark}
+	class="rounded-full p-3 flex place-items-center place-content-center transition-all
 	{isOpacity ? 'hover:bg-immich-bg/30' : 'immich-circle-icon-button hover:dark:text-immich-dark-gray'}
   {forceDark && 'hover:text-black'}"
 	on:click
